Feat: rename folders
