+++ src/main/java/com/assessment/riwi/infrastructure/abstract_services/generics/CreateService.java
@@ -1,4 +1,4 @@
-package com.assessment.riwi.infrastructure.abstract_services.generic;
+package com.assessment.riwi.infrastructure.abstract_services.generics;
 
 public interface CreateService <requestDTO, responseDTO>{
     responseDTO create(requestDTO request);
