+++ src/main/java/com/assessment/riwi/infrastructure/abstract_services/generics/DeleteByIdService.java
@@ -1,4 +1,4 @@
-package com.assessment.riwi.infrastructure.abstract_services.generic;
+package com.assessment.riwi.infrastructure.abstract_services.generics;
 
 public interface DeleteByIdService<ID> {
     void deleteById(ID id);
