+++ src/main/java/com/assessment/riwi/infrastructure/abstract_services/generics/ReadByIdService.java
@@ -1,4 +1,4 @@
-package com.assessment.riwi.infrastructure.abstract_services.generic;
+package com.assessment.riwi.infrastructure.abstract_services.generics;
 
 public interface ReadByIdService<responseDTO, ID> {
     responseDTO readById(ID id);
