+++ src/main/java/com/assessment/riwi/infrastructure/abstract_services/generics/UpdateByIdService.java
@@ -1,4 +1,4 @@
-package com.assessment.riwi.infrastructure.abstract_services.generic;
+package com.assessment.riwi.infrastructure.abstract_services.generics;
 
 public interface UpdateByIdService<requestDTO, responseDTO, ID> {
     responseDTO updateById(ID id, requestDTO request);
